File bug future about bools and forall intents and work around in array.find
Last night's testing revealed a problem with my parallel array.find()
routine for the case of a rectangular array indexed by booleans when
compiled with --baseline.  Simplifying the idiom that it uses down to
a standalone test case and compiling with an older version of the
compiler, I'm finding that this seems to be a pre-existing issue, so
am filing a future for it here to capture the case.

Meanwhile, it occurs to me that an array indexed by bools will likely
never be large enough to warrant parallel execution, so I added some
logic to my conditional selecting between parallel and serial
implementations to fall into the serial implementation for bool
indices (including multidimensional arrays of bools, figuring that
the rank would have to get pretty large before it mattered.

This change makes me think that maybe the enum case should also be
serial since enums are typically small even though they could be large
and explode more correctly in higher-dimensional arrays.  To minimize
changes (and since that argument also applies to arrays that happen to
be over small integer ranges), I'm putting that off rather than doing
more here today.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>