Remove tests of error message when attempting to init= between new types
Now that fileReader and fileWriter are separate types, it seems even less likely
someone will try to assign between the two of them.  We aren't going to add an
`init=` specifically to get these error messages, so the behavior is not really
different than trying to assign between any other disparate types.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>