Add a CHPL_AUTO_MAKE_TARGETS var and 'checkwhitespace' target
This adds a CHPL_AUTO_MAKE_TARGETS Makefile variable that can be used
to create a given target (or targets?) to be made every time you do a
`make all` or `make`.  By setting the variable in your environment or
command-line, you can force every make to run some command.

I also added a 'checkwhitespace' target to Makefile.devel that runs
the same script as our CI process does.  The combination of the two
permits a developer to set CHPL_AUTO_MAKE_TARGETS=checkwhitespace in
order to get a whitespace check on every build (if set in environment,
or every build that sets it on the command-line otherwise).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>