Drop the problem size of some lLockFree{Queue,Stack} futures
The `interleavedTest` futures took ~70s, which we noticed after enabling
futures by default. That's long for a normal test, much less a future so
reduce the problem size. It's small enough now that we can remove `-O3`
from compopts. Note that I did this for all tests in these dirs even
though it was only the interleaved ones that were particularly slow.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>