Git-clone only the current tree of champs
This modifies champs sub_test to clone only the current source tree
without bringing in all branches and dev history. This is done
in order to reduce the cloning time and the size of the cloned tree.
As of this writing, this change reduced the tree size from 499Mb to 7.4Mb.

Signed-off-by: Vassily Litvinov <vasslitvinov@users.noreply.github.com>