Dial day12 AoC entries down to small problem sizes
Yesterday, I was a bit on the fence about whether to include both the
small and large problem sizes for my day 12 entries, or just the small
one, and ended up going with both because it was already implemented
and seemed pretty fast.  Two of the big cases timed out in valgrind,
though, so that was enough to make me kick them out and have this test
the small problem sizes only (which I think were big enough to
demonstrate the bug that Vass recently fixed).

While here, I removed duplicate input and .good files, leaning more
heavily on the .stdin capability for tests.  I also enabled version 3
of the multiversion test since I eliminated so many other cases.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>