Don't name dependent fields in partial instantiation warning
Follow-up to PR #21410.

The goal of this changes in to avoid listing fields that need to be
instantiated when the field has a type that depends on an earlier field,
for example, with 'record R { param p; type eltType; var element: eltType; }',
the warning for 'R(1)' should not include 'element' since it's not
really a generic field (just one that hasn't yet been instantiated).

Instead of actually computing whether such a field is dependent or not,
the approach here is to only report a non-type non-param field with a
type expression if it is the first field. This is an approximation, but
it is easy to implement. If it is wrong, users should see the next
generic field once they add an argument for the reported generic fields.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>