Update spec w.r.t. single-statement 'do'; lack of 'return' special case
This updates the spec's syntax diagrams to add support for
single-statement 'do' subroutine bodies and to remove
the previous exception for single-statement returns.

It also udpates a few examples that I ran into while looking for other
places that might need changing; apparently these are among the
examples that have not been converted to running tests.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>