Failed attempt to get this into better shape
I ended up backing away from the "save and restore module" aspect of
the previous approach because the new error about 'bar' being
unresolved didn't feel attractive to me.  Here, I tried going in and
specializing the error in normalize to complain specifically about
FCFs (on the argument that if we haven't resolved something, it must
be a FCF?  Otherwise, it'd be resolved or we would've complained about
it in scope resolve).  However, as the futures here show, this doesn't
make things much better for most other cases.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>