Document the meaning of CHPL_HOME for `make install` builds
In #21622, it was pointed out that CHPL_HOME is not documented for
users of `make install` builds.  This updates the "Building Chapel"
documentation to indicate that $CHPL_HOME is in the
`share/chapel/x.yz/` directory under the prefix into which Chapel
was installed.

This is not a perfect solution because someone may be using a version
of Chapel installed by someone else and never read those instructions
(and/or they may do the install themselves and never finish reading
the instructions or forget about them).  Still, documenting it here
seems like an obvious step forward.

The main downside to this approach is the use of the `--chpl-home`
flag by the configure script to specify the installation location
since that ends up being a different value than what we're calling
`$CHPL_HOME` here.  That makes me wonder whether we could use a
different flag name for it like `--prefix` and whether there's a
reason we didn't do that from the start...

Another thing we could do is to promote the `--print-chpl-home` to a
user flag (potentially calling it `--print-CHPL_HOME` to improve
chances of finding it).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>