Make `--print-chpl-home` user-facing; add `--print-chpl-loc`
This makes the longstanding `--print-chpl-home` flag user-facing so
that a user who is using a Chapel installation that they didn't build
themselves and read the build docs for can determine the value of
$CHPL_HOME for the sake of understanding error messages and the like.
Though it still requires the user to learn about the flag, since it
will show up in the --help output and man page, it seems more likely
that they'll find it than by reading building.rst/.html.

The old form of `--print-chpl-home` printed CHPL_HOME, a tab, the path
to `chpl`, and then potentially a line indicating the prefix location
if `./configure --prefix` was used prior to building `chpl`.  Here,
I've streamlined `--print-chpl-home` to just print CHPL_HOME and moved
the printing of the compiler's location into its own flag,
`--print-chpl-loc`.

I have not bothered to maintain the printing of the prefix location
because I haven't practically used it, and couldn't find tests or code
that did either.  My thinking is that we can always restore that
if/when needed.

Beyond making the obvious driver / compiler flag changes, I've also
updated the man page and the bash completion script.  An existing
test, `test/compflags/ferguson/home` tested the old
`--print-chpl-home` flag, and has been updated to reflect the new
flags and preserve the same tests it was performing before.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>