Make qthreads config-time check of hwloc linking configurable
Qthreads configuration contained a static list of sets of potential link-time
arguments needed to link against hwloc. The configure script would iterate
through this list until a set of arguments was successful. This was
problematic because it required keeping the list up-to-date with respect to
hwloc changes. This commit makes the check configurable as it is only
beneficial to Chapel developers as any problems with linking against hwloc
will be found when compiling an app. It's not worth the cost of maintaining
the list nor the complexity of adding support for pkg-config.
The --disable-hwloc-configure-checks argument to configure turns off the
hwloc link check. The link check also tests whether or not hwloc supports
distance information. If the link check is disabled, configure assumes that
hwloc supports distance information. This may be changed by
passing --disable-hwloc-has-distance to configure.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>