Deprecate auto-included version of carg and move it to the Math module
We don't want to include this function in all programs.

The commit for this ends up looking a little odd so that we don't accidentally
build the Math module by default as well - make an undocumented version that
Math and the deprecated version can call, so that Math depends on AutoMath and
the two versions stay in sync until the deprecated version is removed.  This
strategy was previously followed for other symbols in the Math module.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>