Deprecate including erf and erfc by default and move them to the Math module
We don't want to include these functions in all programs.

The commit for this ends up looking a little odd so that we don't accidentally
build the Math module by default as well - make undocumented versions that
Math and the deprecated versions can call, so that Math depends on AutoMath and
the two versions stay in sync until the deprecated versions are removed. This
strategy was previously followed for other symbols in the Math module.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>