Enable calling recursive functions from the GPU
When the second instance of a function is detected in a call chain (meaning
it is recursive) return that it is gpuizable. This will enable calling it from
the GPU as long as the rest of the first instance is OK.

The presence of recursive functions makes ptxas unable to statically determine
the stack size, which it warns about by default. Add a flag
"--suppress-stack-size-warning" to the list of flags for ptxas.

Add a test that recursively computes Fibonacci numbers on the GPU.

Update a test that had an infinite recursion to stop after a smaller number
of calls.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>