Add another workaround for LLVM_NODISCARD for intel compilers
This is a follow-on to #20426 that changes another warning to a
diagnostic around LLVM_NODISCARD for the Intel Classic Compiler
2023.1.0 release.  Without it, the GASNet team was seeing:

```
/path/to/chpl-home/third-party/llvm/install/support-only-linux64-x86_64/include/llvm/ADT/StringRef.h(781): error #1292: unknown attribute "clang::warn_unused_result"
      LLVM_NODISCARD
      ^
```

in a developer build (with WARNINGS=1 on).  Note that this is a subtle
shift from the previous warning that didn't seem to recognize the
namespace rather than the attribute itself.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>