Handle reservations correctly in slurm-srun launcher
The slurm-srun launcher uses sinfo to determine the number of cores and the
threads per core on the nodes. The output from sinfo is sorted by the number
of cores so if there are heterogeneous nodes the launcher will limit each
locale to the fewest number of cores per node. However, if there is a slurm
reservation that has fewer cores it will show up in the sinfo output even if
the launcher isn't running in the reservation, causing the launcher to use
the wrong number of cores. This commit fixes that by making the launcher
ignore any sinfo output for reservations that the launcher is not in.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>