Link libChplFrontend statically when statically linking LLVM
This commit is a short-term fix for problems linking 'chpl' with new
uses of llvm and clang by libChplFrontend when using CHPL_LLVM=bundled
(which builds/links LLVM statically). The fix here is to use static
linking for libChplFrontend as well in that event. That avoids tricky
problems that result from mixing static and dynamic linking. In the
future, it would be good to only support dynamic linking of
libChplFrontend as well as LLVM and clang.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>