Fix bug where we weren't generating promoted deprecation or unstable warnings
Because of the way promotion is handled in the production compiler, we were
failing to recognize when a promoted call should have triggered an unstable or
deprecation warning.  Add a check for if these warnings should be generated when
we create the promotion wrapper.

By handling it at promotion wrapper creation, we avoid the problem of generating
too many warnings (and generating warnings for the promotion wrapper name
instead of the name of the original function).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>