Bundle protobuff support into the release
The release documentation has instructions for using it that haven't
worked since it was introduced because we didn't actually bundle the
code into the release.  This PR does so and also adds a testRelease
step to make sure it builds OK.

While here, I made the `make clobber` step for protobuff ignore
failures out of a slight sense of paranoia (given that users have been
bitten by it and reported it) and not seeing any real downside.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>