Stabilize atomic formal and type names
Switch the atomic formal `value` to `val` and change the value type from
`T` to `valType`. `valType` was chosen to match syncs and be similar to
`eltType` on arrays and `idxType`. The formal name `val` was then used
to match `valType` and so we could distinguish the stored "value" in the
atomic and the name of the formal more easily. The `value`->`val` formal
renaming is being done without deprecation since we believe named args
are never used with atomics and if we had a way to disable named args we
may have just used that.

Part of Cray/chapel-private 3730

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>