Remove extra checking in computePointerElementType
It turns out this checking isn't reliable, because the
LLVM 'ptr' value might be a GlobalVariable or an AllocaInst
but where there was an implicit pointer element type conversion
(e.g. if there was a GEP for the 1st element in a global).

So, only use the GlobalVariable or AllocaInst type if the Chapel type is
not available.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>