Update test precomp files to use same C compiler
On systems where 'gcc' defaults to a configuration that is not
position-independent-executables, the clang-15 might be configured to be
position-independent, leading to a linker error when a separately
compiled .o file is linked with a Chapel program.

Note that this is very system dependent. It occurs on a SLES 15 but not
on an Ubuntu 23.04 system.

To fix, just compile these .o files with the same compiler.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>