Eliminate some overhead when invoking serializers
The 'here' call is surprisingly expensive when calling 'read' or 'write'
in a tight loop. The initializer for 'fileReader' and 'fileWriter' that
creates a shallow alias is also a bit expensive. This change no-inits
the alias and manually assigns the values over. 'here' is replaced with
the '_home' variable, which should be the same due to the caller's
on-statement.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>