Update LCALS Timer to to represent clack_t as uint(64)
Previously it was mixing real(64) and uint(64), and the real(64)
one led to the new extern var with wrong type error firing.

Either way, the C standard / POSIX only say that clock_t
needs to be an integer or floating-point type. So it might
be a different type on some platforms.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>