Switch to a -W-based approach for portability and precedent
Daniel pointed out that I was being naive about other compilers
ignoring my GCC pragma, and in doing more research into the root
cause of the problem, I realized that throwing -Wno-stringops-overread
is very similar to the -Wno-stringop-overflow and -Wno-array-bounds
flags we already throw.  So I took that approach instead.

Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>