Skip nightly testing for GPU AMD+AOD tests that are timing out
We have a new test configuration for AMD+AOD testing so it's not unexpected that we have some failing tests.  I want to get things in a stable state where the job is expected to pass and any failure represents a regression.

So for the short-term I'm adding skipif files for these tests.

I'll fork another GH issue to investigate why these tests are failing and decided what our next steps ought to be.

---
Signed-off-by: Andy Stone <stonea@users.noreply.github.com>