Refactor initialization sequence
Refactor the initialization sequence to allow partial initialization of
the topology layer before initialization of the comm layer. In some
configurations the comm layer needs to know an upper bound on the
number of accessible cores, which requires partial initialization of
the topology layer. The precise number of accessible cores isn't known
until after comm layer initialization because the comm layer determines
the number of co-locales on the node, which affects the number of
accessible cores. So we have to split the topology initialization and do
some of it before and some of it after the comm layer.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>