Add co-locale support to -nl argument
Allow the -nl argument to be an expression of the form MxN, where M is the
number of nodes and N is the number of locales per node. N is optional, and
if it is missing the number of locales per node is determined by heuristic,
which currently is to use the value of CHPL_RT_NUM_LOCALES_PER_NODE. For
example, "-nl 4x2" will run 2 locales on each of 4 nodes, for a total of 8
locales. This extended format for the -nl argument only works with those
launchers and platforms that support co-locales, otherwise an error is
reported.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>