Really weird change required to pass one test in paratest
This is really weird, but when using a try...catch (all), this caused
test/errhandling/parallel/forall-wrap-first-error.chpl to fail and
to fail in weird ways, but changing it to a try! fixes the issue.
That seems like (yet another) sign that something is not particularly
strong and stable w.r.t. error handling within forall loops.  Error
mode was also weird:

```diff
> forall-wrap-first-error.chpl:1: In module 'ReThrown':
> forall-wrap-first-error.chpl:3: error: module-scope variable 'count' may be used before it is initialized
> forall-wrap-first-error.chpl:4: note: 'count' declared and initialized here
```

Check out the SHA before this one to reproduce

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>