Add 'useIOSerializers' config param and update fileReader/Writer fields
This commit adds the 'useIOSerializers' config param, which toggles
whether the new serializers implementation is used by default.

This commit also updates the fields of ``fileReader`` and ``fileWriter``
to use class-wrapped serializers as fields, and updates the names of
the type fields to 'serializerType' and 'deserializerType'.

Accessors ``fileReader.deserializer`` and ``fileWriter.serializer`` are
added to allow advanced users to interact with their (De)Serializers.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>