Fix RVF bug when deserializing array type variables
This commit makes the 'do not RVF' pragma available to user code, and
uses it to avoid RVF-ing the return temp in 'fileReader.read'. This
temporary was being RVF'd due to the 'always RVF' property of arrays,
and so we could never PUT the read array back to the originating locale.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>