Update to Vass's definition of aligned; remove safecasts from ':'; update safeCast
I started out here by updating to Vass's definition of how to set
the aligment in a cast, and in doing so wrestled with whether
stride and/or alignment should be using safecasts here.  In doing
so, I discovered that we have a range safeCast method as well, s.t.
it seemed more consistent to have it use safecasts for alignments
and strides and for the normal ':' not to do so.  Then, with my
attention on safeCast, I updated it to more similarly match our
':' implemenation and added a clone of my rangeCasts.chpl test that
tested safeCast casts as well.

In doing that, I realized that we were missing some key safecast
overloads between bools and ints, so added those to Types.chpl
and updated the safeCast documentation to describe the family
of routines.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>