Add a skipif for threadring on GPU locale model
This test hangs at the end of main for the GPU locale model. Add a .skipif
to avoid timeouts for this configuration.

The only differences I see in the AST logs for flat vs. gpu are extra
temp variables to store the domain when creating arrays and some extra wide
references. I'm not sure how either of these cause it to hang on the program
termination lock.

Signed-off-by: David Iten <daviditen@users.noreply.github.com>