Add initial support for heterogeous cores
Some CPUs, such as Intel's Alder Lake, have heterogenous cores that have a
variable number of processing units (threads) and performance. This commit is
an interim solution for handling heterogenous cores until we roll-forward to
hwloc 2.5+ that provides support for distinguishing between types of cores.
In the meantime, the environment variable CHPL_RT_MIN_THREADS_PER_CORE can be
used to indicate that the runtime should only use cores with the specified
number of threads (the default is 1). Thus, if CHPL_RT_MIN_THREADS_PER_CORE=2
on an Alder Lake CPU the runtime will only use the performance cores.

Note that calling locale.numPUs(logical=false, accessible=false) on a CPU with
heterogeneous cores returns an estimate as it cannot determine the number of
processing units in an inaccessible core.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>