Fix problems when 'clang' is not in the path
Follow-up to PR #22454

When working with Nix, we might have a setting described in issue #22490
where clang is not in the PATH but nonetheless various variables are set
to point to it. Before this PR, in such a situation, get_system_llvm_clang
would return '', because 'clang' was available neither in `llvm-config
--bindir` nor from PATH.

The solution is to adjust get_system_llvm_clang to consider the values
of the existing documented configuration variables CHPL_TARGET_CC,
CHPL_TARGET_CXX as well as the undocumented configuration variables
CHPL_LLVM_CLANG_C and CHPL_LLVM_CLANG_CXX.

Note that get_system_llvm_clang can't call chpl_compiler.get('target')
without entering an infinite loop.

Fixes a couple of corner cases I observed during development, as well.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>