Marking Yaml basic-tests.chpl file .notest for tonight
In paratesting this afternoon, I found that this test didn't pass on
chapcs nor on my Mac, and that it hadn't been tested since it was
checked in due to the various outages.  I don't know whether the
output I'm getting is correct or whether something else about the
code should be changed, so rather than trying to fix it or accept
the new output, I'm just marking it a notest for tonight to avoid
unnecessary noise in testing.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>