Resolve a problem using system clang on Alpine linux
We were seeing errors along the lines of

    fatal error: 'stdatomic.h' file not found

These appeared after PR #22454 which changed the printchplenv logic to
compute CHPL_LLVM_CLANG_C etc as 'clang' rather than '/usr/bin/clang'
on this platform. For some reason, using clang as a library without
the full path causes it to miss adding a system include path. So,
this PR just adjusts the clang search process to run 'which' if
it doesn't already have a directory component.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>