Handle CHPL_LLVM_CLANG_C containing arguments
Follow-up to PR #22530. Intended to resolve smoke test failures with
that PR. Certain testing configurations run commands with the
environment including CHPL_LLVM_CLANG_C output by a previous
printchplenv run. As such, it might include arguments, which caused
errors along the lines of 'Error: Missing or wrong version for clang'.

This commit pulls out the consideration of overrides into a helper
function that returns a list and uses the helper function in both
get_system_llvm_clang and get_llvm_clang. The first of these
only uses the 1st list element (to run `clang --version` etc) but the
second returns the full list with the arguments.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>