Address other comments from Engin; introduce sections; more proofreading
This started out as addressing some other minor comments from Engin,
but then that led me into looking at the overall flow of the primer,
finding it lacking, and wanting to introduce sections to make it less
monolithic.  In doing so, I ended up moving or pulling some
computations that felt out-of-place or extraneous in an array primer.
(I left in some other things — like all the content about domains —
despite the fact that it could arguably be subsumed by the domains
primer, mostly since domains and arrays are so tightly coupled).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>