Improve 'canPass' for tuples
Improves 'canInstantiate' to check for passing tuples of mismatched
size, and sets 'isKnownSize' to false for fully-generic tuples (i.e.
"_tuple").

Also updates "CompositeType::areSubsInstantiationOf" to return false if
a generic component doesn't exist in the 'partial' type, which can
happen if this method is invoked with tuples of different sizes.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>