Retire nbody_fullreduction future
First, this commit updates the code to use unmanaged classes so that the
program actually compiles.

To get this program running correctly the assignment to
``B.coord_vector`` in ``advance`` needed to be fixed. This assignment
was incorrectly using a reduction over the vector and assigning all
planets to have the same coord_vector. This commit updates the
assignment to match the logic of the for-loop commented out below the
assignment.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>