Allow user to force creation of fixed heap with comm=ofi
The environment variable CHPL_RT_COMM_OFI_NEED_FIXED_HEAP causes the ofi comm
layer to create a fixed heap even if the provider doesn't require one, or if
there is only one locale (in which case the comm layer isn't fully
initialized). This is useful to applications that want to ensure there is a
fixed heap that can be accessed via chpl_comm_regMemHeapInfo().

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>