Silence unstable output in Beta_Diversity for now
While running a paratest on a branch made from main today, I was
seeing failures in the sum reduce output of the Beta_Diversity test
that was added today, and then testing on my Mac saw a different value
in the output.  This adds a config const for the time being that
suppresses the output so that we'll still get the check that the code
compiles and runs completely, but without any validation that the
output is expected (the goal being to head off potentially getting
a test mail for every configuration tonight if all of them similarly
get different answers).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>