Remove CHPL_RT_COMM_OFI_EP_COUNT
The CHPL_RT_COMM_OFI_EP_COUNT environment variable was added as an expedient
workaround to the erroneous ep_cnt value returned by the cxi provider.
However, it requires documentation and relies on the user to specify a large
enough value to bind all cores to endpoints. Instead, until the cxi bug is
fixed, simply assume the cxi provider can create as many endpoints as we
need.

Signed-off-by: John H. Hartman <jhh67@users.noreply.github.com>