Separate instructions for asan-ing compiler+exe vs. just exe
Rationale: I almost never want to asan the compiler, and find the
traditional instructions that start by setting something and then
unsetting it (while not unsetting other things that need not bet set)
to be roundabout and to require far too much thinking.

I also added a comment to the CHPL_LLVM=none case for exe-only because
there's really no need to rebuild the compiler just to use the C
back-end, so I put the option of just switching to the C back-end as
an option.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>