Fix segfaults in users/engin/sparse_bulk_dist/*.chpl
These were getting segfaults at program teardown in GASNet testing
and a valgrind error in --no-local testing.  Presumably it's
something to do with a class type with generic management style
getting passed in to the sparseLayout field and not being made
into the concrete type 'unmanaged sparseLayout'...  but I'm not
sure what the problem was, nor why it didn't generate some sort
of error (like a mismatch between types or an attempt to allocate
a generic type or something).

Q: If we have `type t = C;`  (i.e., a class with generic management
type) and we do `new t();` do we get an error or a 'new owned t()'?
And, if the latter, how did we ever manage to store that into a
field of type unmanaged t?

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>