Mark atomic compareAndSwap as unstable
We currently have atomic `compareAndSwap` and `compareExchange`.
`compareExchange` matches the C/C++ API and the main difference is that
`compareExchange` takes `expected` by ref and updates it on failure.
This is more efficient when you want to update expected since it avoids
having to do another read to get the value.

However, the API is more cumbersome when you don't want to modify
expected or you're using non-mutable values. We have a fair number of
tests that want the `compareAndSwap` behavior, but we aren't positive
about the name since there's not an intuitive difference between
`compareAndSwap` and `compareExchange` since they're often used
interchangeably. For now mark it unstable until we find a better name or
decide it's good enough as-is.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>