Add a user facing task yield
Explicitly yielding a task is important for avoiding deadlock in
busy-loops and we also use it code like aggregators to give other task a
chance to run and let remote tasks make progress sooner.  However,
historically we have not had a user-facing mechanism and just relied on
the runtime `chpl_task_yield()` function.

This adds a user-facing version `currentTask.yieldExecution()`. This is
implemented as a `currentTask` namespace (similar to C++ `this_thread`
and a `yieldExecution` function. Chapel tasks are anonymous, so this was
done to make it clear you're only operating on the current task and it's
not possible to capture. This namespace is also somewhere we can attach
other task specific functions too in the future (e.g. `currentTask.id`).
We consider task yielding a core part of Chapel, so it's available by
default.

Resolves 10977

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>