Adjust workaround for type constructors
To solve the warning in initeq.chpl, I needed to change it to use `T(?)`
but that caused a compilation error due to a workaround added in PR #13315

Removing the workaround entirely caused failures for 3 tests:

  classes/ferguson/recursive/issue-12612.chpl
  types/records/recursive/generic-simple.chpl
  types/records/vass/issue-11846-a.chpl

So, the workaround seems necessary in some cases today.

This commit adjusts that workaround to be slightly more focused (now it
requires that the parentSymbol is a TypeSymbol) so that the 3 tests
above can use it but it does not interfere with initeq.chpl.

---
Signed-off-by: Michael Ferguson <mppf@users.noreply.github.com>