Cleanup + notes
This works for test/deprecated/new-dmap.chpl but not dmapType.chpl.
The reason is that the signature for the 'Block' type has now
changed from rank, idxType, etc. to the type of the helper record
field.  This seems like a job for user-defined type constructors
(along with user-defined specification of how a type should be
printed out to keep the code view and stringification of the type
in sync with one another), but of course we don't have that yet.

Rewinding, the reason I pushed all the generic stuff into the helper
record was the following: most of the duplicated code refers to fields
like _instance or _pid or _unowned.  But those fields only exist in
the outer record, not the helper record.  So I pushed them into the
helper record.

So, some questions:

* Is there a way to have a helper record forwarding field refer to
  the outer record's fields that I Haven't thought of yet?

* Or maybe I can duplicate the basic, user-facing generic fields in
  the outer and inner records to keep the type signature more
  consistent?  Maybe even completely consistent if I can make the
  helper record no-longer-generically varying between dists?  (which
  may, in itself, be an instance of the kind of "looks generic, but
  it isn't" pattern that I've been worried about lately.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>