Attempt to restore type signature prefixes, if not all
This was my attempt to restore the 'Block(1...)' type signatures
by moving the generic rank, idxType, etc. fields back into Block.
It didn't work because some call chain related to getting the
runtime type of my block distribution uses a rank of 0 (probably
the default for the int field) which causes a problem in the
tuple code.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>