Disable hwloc readme generation
We've historically had issues generating the hwloc README and did some
file touching to try to trick the build system into not generating it.
Hwloc 2.8 added a `--disable-readme` option, so now that we've upgraded
to 2.9 we can use that option instead of touching specific files.

For more context see 19522 and open-mpi/hwloc 526

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>