Remove the deprecated versions of `logBasePow2` from AutoMath.chpl
This function has been deprecated for a few releases and leaving it will cause
conflicts with other deprecations we want to do.  Clean it up (but leave
`chpl_logBasePow2` because other symbols in this module rely on it - add a note
to those functions so we don't forget to move its contents later when they are
moved).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>