Improve isClose error when promotion is involved
In the old version (prior to limiting the types), we would get an error that
mentioned promotion when it was attempted (though this error would reveal
implementation details).  The error with a where clause was significantly less
clear, so adjust the check to occur within the function itself and generate
errors that mention promotion when relevant and otherwise just inform about the
type limitations.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>