Change code implementing atomic max from compareAndSwap to compareExchange
This code wants the old value on failure, so instead of doing another
atomic read to get it, we can just have `compareExchange` update it on
failure (which is the advantage of compareExchange.) This is slightly
more efficient.

Note that some of these are done with read-modify-write semantics and
others are done with read-conditionally-write. Longer term I expect
we'll want atomic min/max in the standard library and we'll want to
unify on one or the other (likely RMW by default to match the C++
proposal.)

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>