Skip tests that use a really small stack size under baseline
Some tests that create a lot of concurrent tasks lower the stack size to
run on systems with limited memory. This was causing issues for a few
tests under baseline since we were exceeded a 128k stack size. This is
an unusual situation (small stack size and baseline) so just skip all
tests using less than 256K stacks under baseline.

Signed-off-by: Elliot Ronaghan <ronawho@gmail.com>